+++ source/ui/RecordUi.cpp
@@ -111,7 +111,7 @@ void RecordUi::showEvent(QShowEvent*)
 }
 void RecordUi::customEvent(QEvent* e)
 {
-	if (e->type() == QiEvent::rec_start) RecStart();
-	else if (e->type() == QiEvent::rec_stop) RecStop();
-	else if (e->type() == QiEvent::rec_close) RecClose();
+	if (e->type() == static_cast<int>(QiEvent::rec_start)) RecStart();
+	else if (e->type() == static_cast<int>(QiEvent::rec_stop)) RecStop();
+	else if (e->type() == static_cast<int>(QiEvent::rec_close)) RecClose();
 }
