+++ source/ui/VarViewUi.cpp
@@ -144,7 +144,7 @@ void VarViewUi::showEvent(QShowEvent*)
 }
 void VarViewUi::customEvent(QEvent* e)
 {
-	if (e->type() == QiEvent::var_reload)
+	if (e->type() == static_cast<int>(QiEvent::var_reload))
 	{
 		if (!isHidden())
 		{
